Fix panic when incbin'ing an empty file
If you incbin an empty file, the code I wrote for including ranges assumes you gave it an invalid start and tries to report an error on an argument that wasn't provided. This fixes that by short-circuiting to returning an empty value if there are no bytes in the file.